Edge change agent emits unenriched types with enriched.
Up until this point, the Edge Change agent emitted only one edge change event
on Relink and Backlink user actions, though, essentially, Relink and Backlink
events are still Link events. Though this is not a bug, this makes it
difficult to reason about counting link events in other agents such as the
rate limiter agent.

For example say you have a limit on Link events, but not Relink events. This
allow users to abuse actions on the site that result in Relinks, despite the
fact that the user is spamming the same action.

Accordingly, this commit changes the edge change transformer to emit a Link
event with every Relink and Backlink event.