Replace DatabaseHealthMonitor with a circuit breaker.
After some consideration, we've decided that a circuit breaker is a better
mechanism to prevent repeated failed requests to an unresponsive database
than an asynchronous health monitor.

If a database is unresponsive, the circuit breaker will declare it inactive
and set a time after which point contacting the database will be retried. The
delay on the retry doubles on each failed attempt.