FIX bug in ring timed lookback.
the following function:

def ordinals(timeLimit: Option[Long], at: Long = now): List[Int]

returned a list with the first bucket given either a lookback of 0 or the full
'ticks' duration. The logic behind this was that, having cycled through the
'ticks' duration, one should return to the first bucket.

This is misleading. Given a lookback equal to the full ticks duration, the
full set of ordinals should be returned. This commit changes and tests this
behaviour of the ordinals function. Now calling

a) ring.ordinals(Some(0L)) will return an empty list.
b) ring.ordinals(Some(ticks)) will return the full set of ordinals.