Add a 'total' field to UserAction.
This field is intended for use in an enrichment agent that would transform a
the user action to a copy of it with the total value modified to the total
number of actions of that class performed by the user.

For example, if an favoriting is ingested into sketchy and the sender has
performed 100 favoritings on the client application, an enrichment agent can
copy the object with the total field set to Some(100) given access to the
client application database.